Create custom_animation.java.ftl
This block allows you to play custom animations on an entity in Minecraft. You can upload your own animation files in the "animations" folder and select them as the animation to play on the entity. To use this block, drag it into your workspace and specify the entity to play the animation on, the animation file to play, and the speed at which to play the animation. If the "animations" folder does not exist, it will be created automatically when the block is added to the workspace. To upload custom animations, simply place the animation files in the "animations" folder and select them in the block.